backend/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "affordabot-backend"
version = "0.1.0"
description = "Affordabot backend service"
authors = ["Feng Ning <fengning@stars-end.ai>"]
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "*"
uvicorn = "*"
pydantic = "*"
python-dotenv = "*"

aiohttp = "*"
litellm = ">=1.0.0"
prefect = ">=2.0.0"
instructor = "*"
openai = "*"
beautifulsoup4 = "*"
httpx = "*"
pypdf = "*"
sqlalchemy = "*"
pyopenstates = "*"
python-dateutil = "*"
six = "*"
sentry-sdk = {extras = ["fastapi"], version = "*"}
sendgrid = "*"
eval_type_backport = "*"
scrapy = "*"
llm-common = {path = "/home/fengning/llm-common", develop = true}
playwright = "^1.56.0"
city-scrapers-core = "^0.10.1"
tenacity = "^8.0.0"
asyncpg = "^0.30.0"
minio = "^7.2.20"
scrapy-playwright = "^0.0.44"
psycopg2-binary = "^2.9.11"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
supabase = "^2.27.0"
pgvector = "^0.4.2"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-asyncio = "*"
black = "*"
ruff = "*"
mypy = "*"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
